requirements.txt
flask
pyjwt
requests
pytest